. Adding implicit casting capabilities to the PostfixExpressionTokenizer . Replaced most (long) casts to long literals in unit tests

+++ ZScriptTests/CodeGeneration/MultipleSourcesTests.cs
@@ -180,7 +180,7 @@ namespace ZScriptTests.CodeGeneration
 
             runtime.CallFunction("f1");
 
-            Assert.AreEqual((long)10, memory.GetVariable("a"), "Cross reference of functions failed to be executed properly");
+            Assert.AreEqual(10L, memory.GetVariable("a"), "Cross reference of functions failed to be executed properly");
         }
 
         [TestMethod]
@@ -201,8 +201,8 @@ namespace ZScriptTests.CodeGeneration
             runtime.CallFunction("f1");
             runtime.CallFunction("f2");
 
-            Assert.AreEqual((long)30, memory.GetVariable("a"), "Cross reference of functions failed to be executed properly");
-            Assert.AreEqual((long)50, memory.GetVariable("b"), "Cross reference of functions failed to be executed properly");
+            Assert.AreEqual(30L, memory.GetVariable("a"), "Cross reference of functions failed to be executed properly");
+            Assert.AreEqual(50L, memory.GetVariable("b"), "Cross reference of functions failed to be executed properly");
         }
 
         #endregion
