+++ ZScriptTests/Runtime/ArrayTests.cs
@@ -52,9 +52,9 @@ namespace ZScriptTests.Runtime
             Assert.IsInstanceOfType(memory.GetVariable("a"), typeof(IList), "The list created by the script is not of the expected type");
             var list = (IList)memory.GetVariable("a");
 
-            Assert.AreEqual((long)0, list[0], "The list did not have the expected values");
-            Assert.AreEqual((long)1, list[1], "The list did not have the expected values");
-            Assert.AreEqual((long)2, list[2], "The list did not have the expected values");
+            Assert.AreEqual(0L, list[0], "The list did not have the expected values");
+            Assert.AreEqual(1L, list[1], "The list did not have the expected values");
+            Assert.AreEqual(2L, list[2], "The list did not have the expected values");
         }
 
         /// <summary>
@@ -80,9 +80,9 @@ namespace ZScriptTests.Runtime
             var list = (IList)memory.GetVariable("a");
 
             Assert.IsInstanceOfType(list[0], typeof(IList), "The list created by the script is not of the expected type");
-            Assert.AreEqual((long)0, ((IList)list[0])[0], "The list did not have the expected values");
-            Assert.AreEqual((long)1, list[1], "The list did not have the expected values");
-            Assert.AreEqual((long)2, list[2], "The list did not have the expected values");
+            Assert.AreEqual(0L, ((IList)list[0])[0], "The list did not have the expected values");
+            Assert.AreEqual(1L, list[1], "The list did not have the expected values");
+            Assert.AreEqual(2L, list[2], "The list did not have the expected values");
         }
 
         /// <summary>
@@ -104,7 +104,7 @@ namespace ZScriptTests.Runtime
             runtime.CallFunction("funca");
 
             // Assert the correct call was made
-            Assert.AreEqual((long)0, memory.GetVariable("a"), "The list did not have the expected values");
+            Assert.AreEqual(0L, memory.GetVariable("a"), "The list did not have the expected values");
         }
 
         /// <summary>
@@ -126,7 +126,7 @@ namespace ZScriptTests.Runtime
             runtime.CallFunction("funca");
 
             // Assert the correct call was made
-            Assert.AreEqual((long)1, memory.GetVariable("a"), "The list did not have the expected values");
+            Assert.AreEqual(1L, memory.GetVariable("a"), "The list did not have the expected values");
         }
 
         /// <summary>
@@ -150,7 +150,7 @@ namespace ZScriptTests.Runtime
             // Assert the correct call was made
             var list = (IList)memory.GetVariable("a");
 
-            Assert.AreEqual((long)1, list[0], "The list created by the script is not of the expected type");
+            Assert.AreEqual(1L, list[0], "The list created by the script is not of the expected type");
         }
 
         /// <summary>
@@ -174,7 +174,7 @@ namespace ZScriptTests.Runtime
             // Assert the correct call was made
             var list = (IList)memory.GetVariable("a");
 
-            Assert.AreEqual((long)2, ((IList)list[0])[1], "The list created by the script is not of the expected type");
+            Assert.AreEqual(2L, ((IList)list[0])[1], "The list created by the script is not of the expected type");
         }
     }
 }
