+++ ZScriptTests/Runtime/ObjectLiteralTests.cs
@@ -65,9 +65,9 @@ namespace ZScriptTests.Runtime
             // Set the dictionary on memory now
             runtime.CallFunction("funca");
 
-            Assert.AreEqual((long)10, ((ZObject)memory.GetVariable("b"))["x"], "The VM failed to correctly create the dynamic object");
-            Assert.AreEqual((long)11, ((ZObject)memory.GetVariable("b"))["abc"], "The VM failed to correctly create the dynamic object");
-            Assert.AreEqual((long)12, ((ZObject)memory.GetVariable("b"))["string with spaces"], "The VM failed to correctly create the dynamic object");
+            Assert.AreEqual(10L, ((ZObject)memory.GetVariable("b"))["x"], "The VM failed to correctly create the dynamic object");
+            Assert.AreEqual(11L, ((ZObject)memory.GetVariable("b"))["abc"], "The VM failed to correctly create the dynamic object");
+            Assert.AreEqual(12L, ((ZObject)memory.GetVariable("b"))["string with spaces"], "The VM failed to correctly create the dynamic object");
         }
 
         [TestMethod]
@@ -86,8 +86,8 @@ namespace ZScriptTests.Runtime
             // Set the dictionary on memory now
             runtime.CallFunction("funca");
 
-            Assert.AreEqual((long)15, ((ZObject)memory.GetVariable("b"))["x"], "The VM failed to correctly create the dynamic object");
-            Assert.AreEqual((long)40, ((ZObject)memory.GetVariable("b"))["y"], "The VM failed to correctly create the dynamic object");
+            Assert.AreEqual(15L, ((ZObject)memory.GetVariable("b"))["x"], "The VM failed to correctly create the dynamic object");
+            Assert.AreEqual(40L, ((ZObject)memory.GetVariable("b"))["y"], "The VM failed to correctly create the dynamic object");
         }
         
         [TestMethod]
@@ -108,8 +108,8 @@ namespace ZScriptTests.Runtime
 
             // TODO: Deal with this absurd nesting of ZObject casts
             Assert.IsInstanceOfType(((ZObject)memory.GetVariable("b"))["a"], typeof(ZObject), "The VM failed to correctly create the nested dynamic objects");
-            Assert.AreEqual((long)10, ((ZObject)((ZObject)memory.GetVariable("b"))["a"])["x"], "The VM failed to correctly create the nested dynamic objects");
-            Assert.AreEqual((long)10, ((ZObject)((ZObject)((ZObject)((ZObject)memory.GetVariable("b"))["b"])["y"])["z"])["a"], "The VM failed to correctly create the nested dynamic objects");
+            Assert.AreEqual(10L, ((ZObject)((ZObject)memory.GetVariable("b"))["a"])["x"], "The VM failed to correctly create the nested dynamic objects");
+            Assert.AreEqual(10L, ((ZObject)((ZObject)((ZObject)((ZObject)memory.GetVariable("b"))["b"])["y"])["z"])["a"], "The VM failed to correctly create the nested dynamic objects");
         }
 
         [TestMethod]
@@ -128,7 +128,7 @@ namespace ZScriptTests.Runtime
             // Set the dictionary on memory now
             runtime.CallFunction("funca");
 
-            Assert.AreEqual((long)10, memory.GetVariable("b"), "The VM failed to perform the correct subscript operation on the ZObject");
+            Assert.AreEqual(10L, memory.GetVariable("b"), "The VM failed to perform the correct subscript operation on the ZObject");
         }
 
         [TestMethod]
@@ -147,7 +147,7 @@ namespace ZScriptTests.Runtime
             // Set the dictionary on memory now
             runtime.CallFunction("funca");
 
-            Assert.AreEqual((long)10, memory.GetVariable("b"), "The VM failed to perform the correct subscript operation on the ZObject");
+            Assert.AreEqual(10L, memory.GetVariable("b"), "The VM failed to perform the correct subscript operation on the ZObject");
         }
         
         [TestMethod]
@@ -166,7 +166,7 @@ namespace ZScriptTests.Runtime
             // Set the dictionary on memory now
             runtime.CallFunction("funca");
 
-            Assert.AreEqual((long)10, memory.GetVariable("b"), "The VM failed to perform the correct member fetch operation on the ZObject");
+            Assert.AreEqual(10L, memory.GetVariable("b"), "The VM failed to perform the correct member fetch operation on the ZObject");
         }
         
         [TestMethod]
@@ -185,7 +185,7 @@ namespace ZScriptTests.Runtime
             // Set the dictionary on memory now
             runtime.CallFunction("funca");
 
-            Assert.AreEqual((long)10, memory.GetVariable("b"), "The VM failed to perform the correct member fetch operation on the ZObject");
+            Assert.AreEqual(10L, memory.GetVariable("b"), "The VM failed to perform the correct member fetch operation on the ZObject");
         }
     }
 }
