+++ ZScriptTests/ZScriptTests.cs
@@ -115,7 +115,7 @@ namespace ZScriptTests
             runtime.CallFunction("f");
 
             Assert.IsTrue(runtime.GlobalMemory.HasVariable("a"), "The variable 'a' should be globally visible after being set by a function that was called through the Runtime");
-            Assert.AreEqual((long)10, runtime.GlobalMemory.GetVariable("a"));
+            Assert.AreEqual(10L, runtime.GlobalMemory.GetVariable("a"));
         }
 
         // Perhaps the only really useful test here, tests parsing of a large script containing all the language constructs available to test.
